+++ include/event/FundsDebited.h
@@ -6,12 +6,12 @@ class FundsDebited: public IDomainEvent {
 public:
     static constexpr std::string EVENT_TYPE = "FundsDebited";
 
-    FundsDebited(std::string event_id, int user_id, int quantity);
+    FundsDebited(std::string event_id, int user_id, double quantity);
     explicit FundsDebited(const std::string &json_string);
 
     std::string toJson() const override;
 
 private:
     int user_id_;
-    int quantity_;
+    double quantity_;
 };
